+++ packages/wells/src/client/model/Cluster.ts
@@ -1,8 +1,0 @@
-export enum Cluster {
-  BP = 'bp',
-  API = 'api',
-  GREENFIELD = 'greenfield',
-  BP_NORTHEUROPE = 'bp-northeurope',
-  AZURE_DEV = 'azure-dev',
-  BLUEFIELD = 'bluefield',
-}